# fastapi-generated/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kong Guard AI API
  version: 1.0.0
  description: |
    Autonomous API Threat Response Agent for Kong Gateway

    This API provides comprehensive threat detection, analysis, and automated remediation
    capabilities for Kong Gateway deployments. It includes real-time threat monitoring,
    AI-powered analysis, incident management, and performance optimization.
  contact:
    name: Kong Guard AI Support
    email: support@kongguard.ai
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.kongguard.ai/v1
    description: Production server
  - url: https://staging.kongguard.ai/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
...
paths:
  /config:
    get:
      summary: Get current configuration
      operationId: getConfiguration
      tags:
        - Configuration
      responses:
        '200':
          description: Current plugin configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PluginConfiguration'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      summary: Update configuration
      operationId: updateConfiguration
      tags:
        - Configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PluginConfiguration'
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    patch:
      summary: Partially update configuration
      operationId: patchConfiguration
      tags:
        - Configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PartialConfiguration'
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigurationResponse'
  /threats:
    get:
      summary: List detected threats
      operationId: listThreats
      tags:
        - Threats
      parameters:
        - name: start_time
          in: query
          schema:
            type: string
            format: date-time
          description: Start time for threat query
        - name: end_time
          in: query
          schema:
            type: string
            format: date-time
          description: End time for threat query
        - name: severity
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
          description: Filter by threat severity
        - name: status
          in: query
          schema:
            type: string
            enum: [detected, analyzing, mitigated, blocked]
          description: Filter by threat status
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of detected threats
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  threats:
                    type: array
                    items:
                      $ref: '#/components/schemas/Threat'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /threats/{threat_id}:
    get:
      summary: Get threat details
      operationId: getThreat
      tags:
        - Threats
      parameters:
        - name: threat_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Threat details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreatDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /threats/{threat_id}/analyze:
    post:
      summary: Trigger AI analysis for threat
      operationId: analyzeThreat
      tags:
        - Threats
      parameters:
        - name: threat_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                model:
                  type: string
                  default: gpt-4o-mini
                  description: AI model to use for analysis
                deep_analysis:
                  type: boolean
                  default: false
                  description: Enable deep analysis mode
      responses:
        '202':
          description: Analysis initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  schemas:
    PluginConfiguration:
      type: object
      properties:
        dry_run_mode:
          type: boolean
          default: true
          description: Enable dry run mode - logs threats but doesn't take action
        threat_threshold:
          type: number
          minimum: 1.0
          maximum: 10.0
          default: 8.0
          description: Threat level threshold for triggering responses
        max_processing_time_ms:
          type: integer
          minimum: 1
          maximum: 100
          default: 5
          description: Maximum processing time per request in milliseconds
        enable_rate_limiting_detection:
          type: boolean
          default: true
        rate_limit_window_seconds:
          type: integer
          minimum: 1
          maximum: 3600
          default: 60
        rate_limit_threshold:
          type: integer
          minimum: 1
          maximum: 10000
          default: 150
        enable_ip_reputation:
          type: boolean
          default: true
        ip_whitelist:
          type: array
          items:
            type: string
          default: []
          description: Whitelisted IP addresses or CIDR blocks
        ip_blacklist:
          type: array
          items:
            type: string
          default: []
          description: Blacklisted IP addresses or CIDR blocks
        enable_payload_analysis:
          type: boolean
          default: true
        max_payload_size:
          type: integer
          minimum: 1024
          maximum: 10485760
          default: 262144
        suspicious_patterns:
          type: array
          items:
            type: string
          default:
            - "union.*select"
            - "drop.*table"
            - "<script"
            - "javascript:"
            - "eval\\("
        ai_gateway_enabled:
          type: boolean
          default: false
        ai_gateway_model:
          type: string
          default: gpt-4o-mini
        ai_analysis_threshold:
          type: number
          minimum: 1.0
          maximum: 10.0
          default: 6.0
        enable_auto_blocking:
          type: boolean
          default: false
        block_duration_seconds:
          type: integer
          minimum: 60
          maximum: 86400
          default: 1800
    PartialConfiguration:
      type: object
      description: Partial configuration update
      additionalProperties: true
    ConfigurationResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
        configuration:
          $ref: '#/components/schemas/PluginConfiguration'
        timestamp:
          type: string
          format: date-time
    Threat:
      type: object
      required:
        - id
        - timestamp
        - type
        - severity
        - source_ip
      properties:
        id:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        type:
          type: string
          enum: [sql_injection, xss, rate_limit, ip_reputation, path_traversal, suspicious_payload]
        severity:
          type: string
          enum: [low, medium, high, critical]
        threat_score:
          type: number
          minimum: 0
          maximum: 10
        source_ip:
          type: string
          oneOf:
            - format: ipv4
            - format: ipv6
          description: Source IP address (IPv4 or IPv6)
        path:
          type: string
        method:
          type: string
          enum: [GET, POST, PUT, DELETE, PATCH, HEAD, OPTIONS]
        status:
          type: string
          enum: [detected, analyzing, mitigated, blocked]
        action_taken:
          type: string
          enum: [logged, rate_limited, blocked, none]
    ThreatDetail:
      allOf:
        - $ref: '#/components/schemas/Threat'
        - type: object
          properties:
            request_headers:
              type: object
              additionalProperties:
                type: string
            request_body:
              type: string
            response_status:
              type: integer
            analysis:
              type: object
              properties:
                ai_analysis:
                  type: string
                confidence:
                  type: number
                  minimum: 0
                  maximum: 1
                patterns_matched:
                  type: array
                  items:
                    type: string
                recommendations:
                  type: array
                  items:
                    type: string
...
    AnalysisResponse:
      type: object
      properties:
        analysis_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        estimated_completion:
          type: string
          format: date-time
        model:
          type: string
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        has_more:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.kongguard.ai/oauth/authorize
          tokenUrl: https://auth.kongguard.ai/oauth/token
          scopes:
            read:config: Read configuration
            write:config: Modify configuration
            read:threats: View threats
            write:threats: Manage threats
            read:incidents: View incidents
            write:incidents: Manage incidents
            admin: Full administrative access
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - OAuth2: []